Unit tests for console/Proper handling of SVM model file
Add unit tests for console. This gets us to the coverage:

__init__.py 100%
text.py 100%
model.py 96%
console.py 86%

Not testsing the Python 2 specific code in model.py. Not testing the
demo code in console.py.

This also loads the correct model type for predict and evaluate
depending on whether we were given a file or a directory.

This resolves #12. This resolves #22.