Support Large Vocabularies
Store the vocabulary size and recreate the vocabulary and embedding
matrices when deserializing. This allows models with very large
vocabularies to be serialized.

If the vocabulary size of a model is not specified on the
command line, determine it from the training data by including
all the types that have an encoding vector in the vocabulary.

Refactor RNN and convolutional models beneath a shared sequential
embedding model base class.

This changes the programatic interface in a non-backcompatible way
so bump the major version number.

This closes https://github.com/wpm/nlp/issues/10.